背景軽量化

+++ 寿司チョコレート-流れ魚/寿司チョコレート-流れ魚/ObjBackground_2.h
@@ -22,7 +22,7 @@ private:
 
 	float m_y1; //背景①の描画情報
 	int count;//画面切り替え用
-	int n;//基本速度変数
+	float n;//基本速度変数
 	bool end_P;//Background1の終端受け取り用
 	bool End_P;//Background1の終端受判定用
 };
